[frontend] Move vizualization components to widgets
