+++ opencti-platform/opencti-front/src/private/components/widgets/WidgetAuditsViz.tsx
@@ -1,26 +1,23 @@
-import React, { memo, ReactNode } from 'react';
-import StixDomainObjectBookmarksList from '@components/common/stix_domain_objects/StixDomainObjectBookmarksList';
-import StixCoreObjectsNumber from '@components/common/stix_core_objects/StixCoreObjectsNumber';
-import StixCoreObjectsList from '@components/common/stix_core_objects/StixCoreObjectsList';
-import StixCoreObjectsDistributionList from '@components/common/stix_core_objects/StixCoreObjectsDistributionList';
-import StixCoreObjectsMultiVerticalBars from '@components/common/stix_core_objects/StixCoreObjectsMultiVerticalBars';
-import StixCoreObjectsMultiLineChart from '@components/common/stix_core_objects/StixCoreObjectsMultiLineChart';
-import StixCoreObjectsMultiAreaChart from '@components/common/stix_core_objects/StixCoreObjectsMultiAreaChart';
-import StixCoreObjectsTimeline from '@components/common/stix_core_objects/StixCoreObjectsTimeline';
-import StixCoreObjectsDonut from '@components/common/stix_core_objects/StixCoreObjectsDonut';
-import StixCoreObjectsPolarArea from '@components/common/stix_core_objects/StixCoreObjectsPolarArea';
-import StixCoreObjectsMultiHorizontalBars from '@components/common/stix_core_objects/StixCoreObjectsMultiHorizontalBars';
-import StixCoreObjectsHorizontalBars from '@components/common/stix_core_objects/StixCoreObjectsHorizontalBars';
-import StixCoreObjectsRadar from '@components/common/stix_core_objects/StixCoreObjectsRadar';
-import StixCoreObjectsMultiHeatMap from '@components/common/stix_core_objects/StixCoreObjectsMultiHeatMap';
-import StixCoreObjectsTreeMap from '@components/common/stix_core_objects/StixCoreObjectsTreeMap';
-import StixCoreObjectsWordCloud from '@components/common/stix_core_objects/StixCoreObjectsWordCloud';
-import type { Widget } from '../../../../utils/widget/widget';
-import { computerRelativeDate, dayStartDate, formatDate } from '../../../../utils/Time';
-import { useRemoveIdAndIncorrectKeysFromFilterGroupObject } from '../../../../utils/filters/filtersUtils';
+import { memo, ReactNode } from 'react';
 import { Box } from '@mui/material';
+import AuditsNumber from '@components/common/audits/AuditsNumber';
+import AuditsList from '@components/common/audits/AuditsList';
+import AuditsDistributionList from '@components/common/audits/AuditsDistributionList';
+import AuditsMultiVerticalBars from '@components/common/audits/AuditsMultiVerticalBars';
+import AuditsMultiLineChart from '@components/common/audits/AuditsMultiLineChart';
+import AuditsMultiAreaChart from '@components/common/audits/AuditsMultiAreaChart';
+import AuditsDonut from '@components/common/audits/AuditsDonut';
+import AuditsPolarArea from '@components/common/audits/AuditsPolarArea';
+import AuditsHorizontalBars from '@components/common/audits/AuditsHorizontalBars';
+import AuditsRadar from '@components/common/audits/AuditsRadar';
+import AuditsMultiHeatMap from '@components/common/audits/AuditsMultiHeatMap';
+import AuditsTreeMap from '@components/common/audits/AuditsTreeMap';
+import AuditsWordCloud from '@components/common/audits/AuditsWordCloud';
+import { computerRelativeDate, dayStartDate, formatDate } from '../../../utils/Time';
+import type { Widget } from '../../../utils/widget/widget';
+import { useRemoveIdAndIncorrectKeysFromFilterGroupObject } from '../../../utils/filters/filtersUtils';
 
-interface DashboardEntitiesVizProps {
+interface WidgetAuditsVizProps {
   widget: Widget;
   popover?: ReactNode;
   config: {
@@ -30,11 +27,11 @@ interface DashboardEntitiesVizProps {
   };
 }
 
-const DashboardEntitiesViz = ({
+const WidgetAuditsViz = ({
   widget,
   popover,
   config,
-}: DashboardEntitiesVizProps) => {
+}: WidgetAuditsVizProps) => {
   const startDate = config.relativeDate
     ? computerRelativeDate(config.relativeDate)
     : config.startDate;
@@ -57,19 +54,9 @@ const DashboardEntitiesViz = ({
   }));
 
   switch (widget.type) {
-    case 'bookmark':
-      return (
-        <StixDomainObjectBookmarksList
-          variant={undefined} // because calling js component in ts
-          dataSelection={dataSelection}
-          parameters={widget.parameters as object} // because calling js component in ts
-          height={undefined} // because calling js component in ts
-          popover={popover}
-        />
-      );
     case 'number':
       return (
-        <StixCoreObjectsNumber
+        <AuditsNumber
           endDate={endDate}
           startDate={startDate}
           dataSelection={dataSelection}
@@ -80,21 +67,19 @@ const DashboardEntitiesViz = ({
       );
     case 'list':
       return (
-        <StixCoreObjectsList
+        <AuditsList
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
-          widgetId={widget.id}
           dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
-          title={undefined} // because calling js component in ts
           popover={popover}
         />
       );
     case 'distribution-list':
       return (
-        <StixCoreObjectsDistributionList
+        <AuditsDistributionList
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -106,7 +91,7 @@ const DashboardEntitiesViz = ({
       );
     case 'vertical-bar':
       return (
-        <StixCoreObjectsMultiVerticalBars
+        <AuditsMultiVerticalBars
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -118,7 +103,7 @@ const DashboardEntitiesViz = ({
       );
     case 'line':
       return (
-        <StixCoreObjectsMultiLineChart
+        <AuditsMultiLineChart
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -130,19 +115,7 @@ const DashboardEntitiesViz = ({
       );
     case 'area':
       return (
-        <StixCoreObjectsMultiAreaChart
-          variant={undefined} // because calling js component in ts
-          endDate={endDate}
-          startDate={startDate}
-          dataSelection={dataSelection}
-          parameters={widget.parameters as object} // because calling js component in ts
-          height={undefined} // because calling js component in ts
-          popover={popover}
-        />
-      );
-    case 'timeline':
-      return (
-        <StixCoreObjectsTimeline
+        <AuditsMultiAreaChart
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -154,46 +127,31 @@ const DashboardEntitiesViz = ({
       );
     case 'donut':
       return (
-        <StixCoreObjectsDonut
+        <AuditsDonut
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
           dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
-          height={undefined}
+          height={undefined} // because calling js component in ts
           popover={popover}
         />
       );
     case 'polar-area':
       return (
-        <StixCoreObjectsPolarArea
+        <AuditsPolarArea
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
           dataSelection={dataSelection}
-          parameters={widget.parameters}
+          parameters={widget.parameters as object} // because calling js component in ts
+          height={undefined} // because calling js component in ts
           popover={popover}
         />
       );
     case 'horizontal-bar':
-      if (
-        dataSelection.length > 1
-        && dataSelection[0].attribute?.endsWith('_id')
-      ) {
-        return (
-          <StixCoreObjectsMultiHorizontalBars
-            variant={undefined} // because calling js component in ts
-            endDate={endDate}
-            startDate={startDate}
-            dataSelection={dataSelection}
-            parameters={widget.parameters as object} // because calling js component in ts
-            height={undefined} // because calling js component in ts
-            popover={popover}
-          />
-        );
-      }
       return (
-        <StixCoreObjectsHorizontalBars
+        <AuditsHorizontalBars
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -205,7 +163,7 @@ const DashboardEntitiesViz = ({
       );
     case 'radar':
       return (
-        <StixCoreObjectsRadar
+        <AuditsRadar
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -217,7 +175,7 @@ const DashboardEntitiesViz = ({
       );
     case 'heatmap':
       return (
-        <StixCoreObjectsMultiHeatMap
+        <AuditsMultiHeatMap
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -229,7 +187,7 @@ const DashboardEntitiesViz = ({
       );
     case 'tree':
       return (
-        <StixCoreObjectsTreeMap
+        <AuditsTreeMap
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
@@ -241,13 +199,13 @@ const DashboardEntitiesViz = ({
       );
     case 'wordcloud':
       return (
-        <StixCoreObjectsWordCloud
+        <AuditsWordCloud
           variant={undefined} // because calling js component in ts
           endDate={endDate}
           startDate={startDate}
           dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
-          height={undefined}
+          height={undefined} // because calling js component in ts
           popover={popover}
         />
       );
@@ -256,4 +214,4 @@ const DashboardEntitiesViz = ({
   }
 };
 
-export default memo(DashboardEntitiesViz);
+export default memo(WidgetAuditsViz);
